Add some more value tests
Adds tests for floating points with exponents, escaped chars and
strings and array class literals to improve coverage.

It was necessary to switch the IntelliJ Idea java version from 17 to 21
to match the version used by gradle due to the latter having a bug fix
which affects the formatting of some floating point numbers which
caused tests to pass in IntelliJ but fail in gradlew.

Bug: 354633349
Test: ./gradlew
Change-Id: Ia88c7361c64713715a089043c0dee648134646de