Move caching into runCodebaseProducerTest(...)
Previously, caching was done when the `TestFile`s were created which
complicated creation as it had to remember to cache the files. It also
coupled the `TestFile`s with the `testFileCacheRule`. This change moves
the caching in to `runCodebaseProducerTest(...)` which simplifies the
caching. It does mean that the same files will be cached multiple times
but the cache will ensure that they all end up using the same cached
file.

This also simplifies the follow-up change which breaks the test into
multiple separate test classes as it is taking too long to run.

Bug: 354633349
Test: ./gradlew
Change-Id: If38d80c5f3ab6921070c30c8d90dcfc17fa6f395