Build jar file for each ValueUseSite test
Previously, each subclass of `BaseCommonParameterizedValueTest` that
tested a specific `ValueUseSite` would create and cache their own copy
of the jar file for all the `TestClass`es used by all of them. This
change creates a jar file specific to each subclass that only uses the
files used by that subclass.

Bug: 354633349
Test: ./gradlew
Change-Id: I0dd99f167db7aa77ddf7ec6bcb4f4f44a8e53258