Dedup the test logic in BaseCommonParameterizedValueTest subclasses
Previously, each subclass would get the actual value, get the
expectation and then compare the two. The first part is specific to the
subclass but the other is not, although it does depend on the
`ValueUseSite` being tested by the subclass. This change extracts the
first part into a `legacySourceGetter` lambda and passes that along
with the `ValueUseSite` to `BaseCommonParameterizedValueTest` and then
uses that to implement the test.

Bug: 354633349
Test: ./gradlew
Change-Id: Ic16b6201d6c41ea0f979db88c6d3ebc288381a5f