Add ArrayValue and ArrayElementValue
`Value`s do not support nested arrays, so the model prevents them from
being constructed, by requiring an `ArrayValue` to contain
`ArrayElementValue` (not `Value`) but `ArrayValue` is not an
`ArrayElementValue`.

`Value`s also do not support arrays of different kinds of values (at
least not yet) so this prevents them from being constructed. In future
it will be possible to mix constant fields with `ConstantValue`s in the
same `ArrayValue` as long as the constant field's value is the same
kind as `ConstantValue`s.

Bug: 354633349
Test: ./gradlew
Change-Id: I3e48a3c55514321a95fe01806004535c222cd1fd