Track ConversionExample creation
Makes use of `EntryPointCallerTracker` to track the creation of
`ConversionExample`s in `ParameterizedPrimitiveValueConversionTest`.
It was necessary to replace `UNSUPPORTED` val with an `unsupported()`
method so that each use of it will create its own `ConversionExample`
to be tracked.

Will make development of follow-up changes which add new examples
quicker to implement.

Bug: 354633349
Test: ./gradlew
Change-Id: Ida389aef64d731a6c091f417b5212c4c5788d47c