Use ValueParser in AnnotationItems
Previously, `AnnotationItem`s used `ValueProvider.fromText(...)` but
that was a placeholder that always threw an exception. This change
replaces it with `ValueParser`, modifying it to implement
`ImplementationValueToModelFactory`.

In order to create a `ValueProvider` for `AnnotationAttribute`s it is
necessary to provide the `AnnotationItem` so the
`CachingAnnotationValueProvider` can use it to try and resolve the
attribute's `TypeItem`. Unfortunately, there are two cases where the
`AnnotationItem` cannot be provided.
1. In `AnnotationsMerger` which creates `AnnotationAttribute`s in order
   to create the `AnnotationItem` using
   `Codebase.createAnnotationFromAttributes(...)`.
2. In `AnnotationFilter` which creates `AnnotationAttribute`s without
   creating an `AnnotationItem` at all.

In both case the `Value` is not currently used, so it is safe to use
the `ValueProvider.UNSUPPORTED`.

Bug: 354633349
Test: ./gradlew
Change-Id: I163fd4083a0e9f2d1d5b6c664766509bebf84474