Make checkExpectedValue(...) conditionally strict
Previously, `checkExpectedValue(...)` was lax as it would always ignore
`ValueProviderException`s. That meant it would be possible for
regressions to easily be introduced. This change prevents that by
making the strictness dependent on the `ValueKind` of the
`expectedValue`. That way, once a `ValueKind` is supported on all three
model implementations it can add it the set of `ValueKind`s which
should do strict testing at which point any regressions will be
detected.

Bug: 354633349
Test: ./gradlew
Change-Id: I6c02ae7751a9526047b986877cf7f3781c3dff85