Stop using AnnotationItem.toString() in assertAttribute(...)
Previously, the `AnnotationItem.assertAttribute(...)` extension
function would always use `AnnotationItem.toString()` to construct the
message, even if the assertion did not fail. That causes problems when
adding class literal tests as using class literal in an annotation
attribute value in Kotlin throws an exception causing the
`assertAttribute(...)` function to always fail.

The intent behind using it was to show the existing attributes in the
annotation. This change replaces the use of the `toString()` method
with the `qualifiedName` property and adds a list of attribute names
instead.

Bug: 354633349
Test: ./gradlew
Change-Id: I54534eb20a79af38458d1f7bce3396382d74572b