Fix class literals in UAnnotationItem.toSource()
Previously, `UAnnotationItem.toSource()` would throw a class cast
exception when run on a class literal expression, e.g. `String::class`.
The `Value` work is not intending to change the output of the existing
legacy code until after migrating it all to `Value` to avoid causing
churn in signature files. However, as this is fixing a class cast
exception there is no chance that this will affect the signature files.

This change fixes it now to reduce future churn in baseline files. The
implementation is not a particularly good one as it does not fully
qualify the types but it does have the same behavior as the other
legacy value code which should simplify migation to `Value`.

Bug: 354633349
Test: ./gradlew
Change-Id: Ia22586edbf95c73d300a2e8c1fe03106606fb41c