Refactor SignatureErrorReporter
Previously, it was used mainly for reporting issues with types but was
also used for reporting one issue with signature files. This change
extracts a method for reporting in `ApiFile` and replaces the latter
use of `SignatureErrorReporter` with a direct call to that method. It
then renames `SignatureErrorReporter` to `TypeItemParserErrorReporter`
and adds `Issues.TYPE_PARSE_ERROR`.

This is preparation for providing a `TypeItemParser` in the model for
use by `ValueParser` for parsing class iteral value, e.g.
`int[].class`.

Bug: 354633349
Test: ./gradlew
Change-Id: Ied9fd53ddd69525d558caeee0204c1edbaa8eab9