Use qualified classes for ValueExamples in signature files
Previously, the `ValueExample`s that imported classes in order to test
unqualified classes were broken on the text model as it requires all
classes to be qualified (unless they come from `java.lang`). This
change adds `signatureExpression` to make sure it uses qualified
classes. That does break some of the legacy tests but they are not so
important for signature files.

This also makes `ParameterizedValueParserTest` use
`signatureExpression` instead of `javaExpression`.

Bug: 354633349
Test: ./gradlew
Change-Id: I1fdc097b6e5040daa78fd35986f9866616034f96